Refresh the ring after activating a bucket type
This is a somewhat heavy hammer, but it is expected that bucket types
are created somewhat infrequently. It is required to notify other
applications on top of riak core that may need this changing
information. In particular it's required by riak_kv to bootstrap riak
ensembles.

Supercedes #617
Part of a fix for basho/riak_kv#1007
Requires basho/riak_kv#1008