Alert riak_kv_ensembles server of possible change
The riak_kv_ensembles gen_server determines which ensembles need to be
created based on the current ring and n_vals corresponding to both
existing buckets and active bucket types. It only runs this check on
ring changes though. We need to alert it to check again in case a
strongly consistent bucket type was activated.

Part of a fix for basho/riak_kv#1007
Requires https://github.com/basho/riak_kv/pull/1008