provide behaviour info for riak_core_coverage_fsm with -callbacks
There are three modules in riak_kv implementing this behaviour:
riak_kv_{buckets,keys,index}_fsm.  The wirst two are good, but the
"index" one has process_results/2 deprecated (L191 has this comment:
"Legacy, unsorted 2i, should remove?"), instead exporting process_result/3,
which is not a legislated callback of that behaviour.

A quick resolution to this, somewhat nitpicky, issue would be to just let
riak_kv_index_fsm not implement this behaviour.  Hereby I'm soliciting
more informative suggestions -- or we could just let it be, along with a
corresponding entry in riak_kv/dialyzer.ignore-warnings, and wait for a
more sweeping change to flush this tiny piece of impropriety.