Fix for heavily overloaded vnode proxies that stay stuck in overload after recovery until possibly thousands of messages are sent.
As the mailbox is only an estimate, simplify by not including the
vnode proxy ping/pong in it - there is no need.  The ping/pong
mechanism will now just track proxied requests being completed.

Previously the mailbox estimate grew incorrectly when in the overload
state.  Now the estimate is only updated when messages are proxied.

If the proxy has to perform a hard check on the msgq len, *do*
include the ping message as part of it so that the estimate will
be correct once the ping/pong happens.

QuickCheck test added for riak_test to exercise more than the
unit tests.