Hashtree fixes for issues identified while debugging yz_aae_test
multi_select_segment / iterator do not handle a couple of edge cases
that occur when an explicitly listed segment is absent (i.e. after the
last entry in that segment is deleted).  In which case iterate returns to multi_select_segment and it does not process the left over remaining_segments with empty lists.

The test found an additional issue with seeking after prefetch duplicating the entries in the final accumulator.

Both bugs are fixed and some additional logic was added to catch the case where the length of dirty hashes list does not match the dirty segments list.  If that happens, the code will log and revert to a full rebuild.