First pass at cuttlefish support for setting async job accept/reject criteria.
This populates the app variable checked by riak_core_util:job_class_enabled/1.

Schema files in repos that want to set criteria do so with mappings as described in the comment.

The schema's syntax has been validated, but its behavior has not been - checking it in as a starting point subject to change.