Add support for batch registration
Recently Andy Till identified a problem when downgrading a node after
upgrading to a version of Riak with an entirely new capability. The new
capability was retained in the ring structure, and since the old version
of Riak didn't know about the capability to change the value, other
nodes assumed the downgraded node retained the feature.

The fix to this problem was to filter out unknown capabilities.

However, since capabilities are registered one at a time and filtering
occurs after each registration and during scheduled ticks, correct
capabilities could be lost.

The only readily-apparent way to reconcile these problems is to register
all capabilities simultaneously. The capability server will be blocked
while the registrations occur, so the tick cannot cause a race
condition, and the filtering that occurs during `update_supported/2`
will be applied after the batch has been recognized.

One feature *not* supported by this change is the legacy support that
was intended to allow capabilities-savvy nodes to negotiate with older
Riak nodes that pre-dated capabilities entirely, relying on environment
variables instead. This feature should be dropped entirely as it is no
longer relevant to modern versions of Riak.