Rename all occurences of source
The name "source" was a bit confusing, so it has been renamed
"provider" (as a binding name) or "service_provider" (as a record
field).

Also corrected the 3rd element of the invocation tuple: it's not a list
of terms, it's just a term (which, of course, can be a list).