riak_core_info_service (#898)
* PoC for information service solution.

riak_core_info_service*.erl files contains the solution.
* Make the info service process do the riak_core call

Changed the code placement of where the result to give back to the dependent process happens.
Now the Result is calculated in the riak_core_info_service_process callback_router function and then the handler is invoked.
* Cleanup service. Hard code eleveldb process for right now.
* Export `callback_router/5` because otherwise `spawn` doesn't work.
* More cleanup...
- Add `state` record to `riak_core_info_service_process` to resolve issues with order of parameters.
- Make children of `riak_core_info_service_sup` `permanent` rather than `temporary`
* Add "TODO" in riak_core_app to move some code to `riak_kv` later.
* Fix docs in `riak_core_info_service` to match final implementation details.
- Also, rename `get` handler to `invoke`
* remove logging statement
* Dialyzer fixes and additional sys:handle_debug call to also call
`handle_debug` on outbound messages.
* Dialyzer fixes.
* Make `riak_core_info_service_process` a `gen_server`.
- Invoke the `shutdown` callback in `Mod:terminate`
* Create `state` record in `start_link` so we can just pass it to `init`
rather than passing all the parameters down to `init`
* Add standard license headers

Also picked up a few spurious whitespace fixes.
* Add edoc cross-references

Since `riak_core_info_service` provides such a detailed doc string,
refer to it from the `_process` and `_sup` modules.
* Extensive edoc updates

* Add paragraph breaks
* Add section headers
* Add cross-references
* Add code markup
* Add type documentation for `callback()`
* Clarify usage docs
* Rename all occurences of source

The name "source" was a bit confusing, so it has been renamed
"provider" (as a binding name) or "service_provider" (as a record
field).

Also corrected the 3rd element of the invocation tuple: it's not a list
of terms, it's just a term (which, of course, can be a list).
* Edoc bugfix

Unrelated to this branch but required to get `make docs` to work: move a
< sign into a code span so that it doesn't throw off the XHTML parsing.
* Add unit tests
* Message handling updates per code review

* Use proper OTP shutdown mechanism when `callback_shutdown` message is
  received.
* Log and shutdown gracefully when an unknown message is received.
* Add error handling

Impose more structure:

* Callbacks can no longer be `undefined`
* The primary consumer callbacks (registration, response handler) must
  return `ok` or the service process will terminate
* Catch exceptions
* Log exceptions and other unexpected return values
* Add "Error handling" section to module doc
* Fix code_change

As Ted noted, must return an `ok` tuple.
* Address type problems

* Explicitly disregard the return value from `apply_callback/2`
* Fix the error tuple typespec for `apply_callback/2`
* Register a shutdown handler

The eleveldb callbacks in this file did not previously include a
shutdown registration. It seems self-evident that if the service process
dies prematurely we want to spin up a new one, so register the
`start_eleveldb_info_service/0` function from `riak_core_app` as its own
shutdown handler.

Presumably in the case of an application/node shutdown the service
supervisor would terminate preventing this from turning into an infinite
loop, but I do wonder whether there should be some differentiation for
application shutdown vs process failure.
* Add end-to-end unit test

Spin up the supervisor and ask `riak_core_ring` for a fresh ring.

Also a minor doc update to clarify the provider parameters passed in the result.
* Add unhappy path test

Add a test which throws an exception when receiving a response from
riak_core and make sure we see a shutdown message.
* Clean up tests

These tests were failing when invoked as part of a larger suite due to
unrelated messages arriving from earlier tests.

* Flush the mailbox during setup
* Be more selective about which messages we process
* Add more comments
* Refactor away some boilerplate redundancy
* Fix eleveldb info service shutdown

* The shutdown callback always takes the info service process pid being
  shut down as an argument, so create a new function to take that pid
* Tell eleveldb the pid is no longer in service (corresponding function is
  already in the appropriate branch for review, MvM is working on the NIF)
* The supervisor will restart the process with the callbacks, including
  registration, so don't try to re-register after shutdown
* Remove unnecessary layer of indirection at shutdown