De-fragilify the service process
The Erlang supervision model does not allow for a child to be removed if
it restarts more often than the intensity/period values permit. Instead,
the supervisor itself and all children are terminated.

This makes the "subscribers" to `riak_core` data susceptible to
completely unrelated errors.

To reduce the probability of this occurring, log non-`ok` return values
from the response handler callback instead of terminating the service
process.