Add participate_in_2i_coverage riak option (#917)
Allows marking of nodes 'down' in calculations for 2i coverage
queries.  In create_plan, it queries each node for the value of the
participate_in_2i_coverage and treats those with the value set
explicitly to `false` as 'down' for the calculation of avaiable
vnodes.

See related riak_kv commit that adds a default value to the KV schema.
This is an operator feature. The operator must bring down the node,
change the config, and start the node, for it to have effect.