Alter seed generation.
Change the seed to use a seed generated based on the default value in riak_core_gossip.erl (given, that's the only callee from this module.)  Ensure that the format provided from the claimaint is the same as the value provided by the default seed generation in riak_core_gossip.  Ensure this format, as gossip will call rand:uniform_s/2 on it, and if that's provided with a normal timestamp in OTP > 20, it must be made using a valid generator.