Introduce last gasp to ring
Allow ring to have metadata indicating it is lastgasp - so that subscribers to ring changes cna ignore the last gasp ring defined just before the close.

Prompt for a tick before the close of the riak_core_claimant - to make sure there is always a tick before the last gasp ring is created.