Don't start vnodes on last gasp
The riak_core_vnode_manager has a mangement tick that may occur during shutdown - that is whilst dependencies are closing, and cor eis not yet closed, the riak_core_vnode_manager could receive and act on a management tick.

When a node is exiting, this cna cause the vnode_manager to start the vnodes required in the fresh ring that was generated as the "last gasp" before closing (so that at on restarting it will be an empty and detatched vnode).

This change prevents that restart - and avoids the associated log noise when those started vnodes fail to start due to the shutdown process.