Tidy stat
The exometer:ensure function requires a list of datapoint types, not just a single type.  The code appears to handle without the list, but higher OTP version dialyzer will complain.

Also an updated version of bear is required to avoid the warning of missing app librariries.  Rather than forking the whole stack of dependencies leading to bear - pulling in a higher version in here.    An alterntaive is fork due to the issues of divergence between the basho and upstream fork.