Add comments
Reviewed previous code, to ensure this is being handled correctly.

However, this does appear to be incorrect - as the current simple_transfer does not handle tail violations.  It only checks forward to the end of the list for TargetN violation - it does not wrap around to the start of the list.