Further refactor
There appears to be a significant issue with the old code.  It chooses at random an offset 0, 1, 2 ... but always starts from the front of the ring.

This means that if ring_size is 8, you will at random get 1 of 3 plans:

[{0, [5, 6, 7]}, {3, [0, 1, 2]}, {6, [3, 4]}]
[{1, [0, 6, 7]}, {4, [1, 2, 3]}, {7, [4, 5]}]
[{2, [0, 1, 7]}, {5, [2, 3, 4]}, {0, [5, 6]}]

But vnode 0 appears in 2 out of 3 plans.  So vnodes 1 to 7 will only be used 1/3rd of the time, but vnode 0 will be used 2/3 of the time - creating a hot vnode.

This refactored code uses the reqid to start at a random point in the ring - then testing every 3rd node (followed by every 3rd node offset by 1 etc).

This now means there are many more coverage plans - and now there is an even spread of utilisation of vnodes across those plans.  See `changing_repeated_vnode_refactor-test/0`.

Refactoring has also further reduced the query planning time.