Mas i1121 passringdir (#992)
* Allow use of ring update functions without app

The riak_kv_console:reip/1 function currently requires the node to be down but riak_core configuration to be available so that functions that rely on get_env will work.

Want to allow for a riak_kv_console:reip/3 where the Ring directory and ClusterName can be passed in.

* Update riak_core_ring_manager.erl

Make Cluster a function arg

* Remove lastgasp check if do_write_rigfile/2

check_lastgasp/1 is not required when node is down (e.g. when running riak_kv_console:reip_manual/1).