Second attempt to manage location_change flag in claim
My understanding is that riak_core_ring:has_location_changed/1 is designed to allow for a re-compute of the ring to be prompted when a location assignment of a node has been changed, but nothing has happened to prompt wants etc.

This is now explicit that we only prompt a claim because of has_location_changed when there exists no Wants that would otherwise prompt a change.  This prevents the choose_claim function from being called ordinarily for a node with no wants (and indeed negative wants), when the location has changed if there are nodes with genuine wants.