Sort claiming members before claiming
The claiming members with the least number of partitions, firstly in its location, then holding itself - should be the first to be passed in to the claim_choose algorithm.

This avoids the issue with a large ring size, of a large time being spent on prefilter_violations in choose_claim ... as every partition has already been claimed.