Change to filter local indexes before filtering for spacing
The prefilter_violations/6 function could take 5 - 10s for large ring sizes, if the node being checked was in the same location as the node with all the current ownership.

All the indices would then need to be checked that they were spaced_by_n.  Now only those indices which are violations, but not already locally assigned are checked.